# speed up company-etags