See #903 

I rewrite the algorithm from the scratch, now the code is much more efficient.

Test with latest linux kernel code.

- The initial loading takes 9 seconds to read tags file and prepare all the candidate. The old algorithm takes more than 120 seconds.

- After initial loading, if CLI program `diff` exists, the new tag name is extracted from diff output. It's less than 0.5 second. The old code need 10~20 seconds at best, or 120 seconds at worst

- `(all-completion prefix all-tags)` is as 50 times fast as old version.