@redguardtoo 2019-07-27
`make test` passed at my computer. Don't know why CI failed.