@cireu 2019-07-29
One test failed on Archlinux, with Emacs 27.0.5 https://github.com/emacs-mirror/emacs/commit/8e0ebb9a3cb9beef2f5ff50436fef1c54a3e3c92

![图片](https://user-images.githubusercontent.com/18224684/62061274-68b47080-b259-11e9-8213-bebfadd753e1.png)