@cireu 2019-07-30
This test also passed on my laptop in Emacs 26.2.