@ShuguangSun 2019-07-31
In travis log, the ```emacs --version``` raise error: can't find libguntls.so...