@dgutov 2019-08-09
Thank you!

I'm happy to review it. Ideally, though, I'd like to see this improvement split into several steps. If you could provide the performance numbers for every change, that would be fantastic.