@redguardtoo 2019-08-10
I just update screencast, start from new Emacs, using linux kernel master. There is timer at right side, so you can see the difference.

Old version:
![old](https://user-images.githubusercontent.com/184553/62818959-1b59cc80-bb92-11e9-9987-ac8b1c349f05.gif)

New version:
![new](https://user-images.githubusercontent.com/184553/62818964-257bcb00-bb92-11e9-86c1-ff2cd41fa0d2.gif)