@dgutov 2019-09-07
Sorry for the late reply.

To merge this I need to understand these things:

- Why these changes can't go into Emacs core instead, to benefit everyone.

- If they can't, then for each significant piece of complexity, the performance benefits and tradeoffs that come with it.