@redguardtoo 2019-09-07
Sure. I can explain.

The reason is simple.  The new algorithm is designed for `company-mode` ONLY. So Emacs core can't use my code.

For example `etags-tags-completion-table` is expected to return a list of strings. But my function returns a complex lisp data structure because of the partition thing.

I could send a new patch to Emacs core team. It's just ten lines of code and is different from my current algorithm.

But in my code, similar function should return