@dgutov 2019-09-08
> But my function returns a complex lisp data structure because of the partition thing

So let's tackle this one first. What performance benefits does this bring? It surely can't speed up parsing. It might even slow it down somewhat. But it might speed up completion. So what are the numbers for both?

> The difference is I use string manipulation API

Did we compare the two directly? I imagine the difference might be something else: like the regular expressions used, or the lack of updating the progress bar.

> I avoid parsing function at all when tags file is updated. I use cli program diff. So the update time decreases from 2 minutes to 0.1 second.

Like discussed, it's an interesting idea, but it should fit within Emacs core, if the other blockers don't stop this from happening.