@redguardtoo 2019-09-08
> > But my function returns a complex lisp data structure because of the partition thing
> 
> So let's tackle this one first. What performance benefits does this bring? It surely can't speed up parsing. It might even slow it down somewhat. But it might speed up completion. So what are the numbers for both?

The complex data object is to speed up searching completion candidates. The data structure returned has no impact on the performance of parsing function. You didn't ask me to test the "searching completion candidates" performance. So all my tests are testing the the process of extract all tag names from tags file content.

See https://github.com/company-mode/company-mode/issues/903#issuecomment-521234754 

All the measurement is to analyzed the performance of extract the completion list from tags file content. Original time is 2 minutes, not it's about 9~10 seconds.

> 
> > The difference is I use string manipulation API
> 
> Did we compare the two directly? I imagine the difference might be something else: like the regular expressions used, or the lack of updating the progress bar.
> 
Yes, see the link above. There is also another test below the link which compares the different regex. The etags-compatible regex costs about 30 seconds while ctags-compatible regex costs 10 seconds. Using string api only.

The original code in Emacs costs about 2 minutes, using buffer API, etags compatible regex.

I did some measurement on progress bar before. It turns out progress bar is not the key element of performance. If you want to I can show the test code and result.

> > I avoid parsing function at all when tags file is updated. I use cli program diff. So the update time decreases from 2 minutes to 0.1 second.
> 
> Like discussed, it's an interesting idea, but it should fit within Emacs core, if the other blockers don't stop this from happening.
The code is written for company-etags only. It has no effect on Emacs core or any other commands using Emacs core API. What does "fit within Emacs core" mean? What are the other "blockers"? Sorry, I'm not native speaker, so I'm a bit confused.