@redguardtoo 2019-09-08
@dgutov, let me explain the original company-etags and what's the bottle neck.

Possible actions:
- a1, read tags file (2s)
- a2, extract all the tag name (2m)
- a3, filter the tag names using prefix (2s)

Scenario a, when company-etags is used for the first time, below actions happen,
- a1
- a2

Scenario b, when user input some prefix, below actions are triggered
- a1 (optional, when tags file is updated)
- a2 (optional, when tags file is updated)
- a3

My optimization:
- I reduce the time of a2 from 2m to 10s in Scenario1
- In Scenario 2, I make a1 and a2 unnecessary by using `diff` (Diff only takes about 1~2 seconds)
- In Scenario 2, I reduce time of a3 from 2s to 0.1s